Update README.md
It's a basic game made with the help of PYTHON by Swapnil Pandey. Player controls deer in left and right direction at the bottom of the screen and two types of entities i.e Hunter and food coming down randomly. Player's motive is to save deer from hunter and give more and more food to the deer.